requirements.txt
discord.py>=2.3.2
aiohttp>=3.9
python-dotenv>=1.0